Completed support for long indices, and added many new tests

+++ src/it/unimi/dsi/sux4j/mph/HollowTrie.java
@@ -39,7 +39,7 @@ public class HollowTrie {
 		
 		CharSequence s = iterator.next();
 		LongArrayBitVector prev = seq2bv( s ), curr;
-		int prefix;
+		long prefix;
 		Node p;
 		
 		while( iterator.hasNext() ) {
